Add Read privilege to Researcher role when granting default privileges.

+++ software/cananolab-webapp/src/gov/nih/nci/cananolab/security/service/AclOperationServiceImpl.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.acls.domain.BasePermission;
 import org.springframework.security.acls.domain.GrantedAuthoritySid;
 import org.springframework.security.acls.domain.ObjectIdentityImpl;
 import org.springframework.security.acls.domain.PrincipalSid;
@@ -193,7 +192,9 @@ public class AclOperationServiceImpl implements AclOperationService
 						aclSid = pSid.getPrincipal();
 					}
 					if (!aclSid.equalsIgnoreCase(CaNanoRoleEnum.ROLE_ANONYMOUS.toString()) &&
-						!aclSid.equalsIgnoreCase(CaNanoRoleEnum.ROLE_CURATOR.toString()) && !aclSid.equalsIgnoreCase(owner)) 
+						!aclSid.equalsIgnoreCase(CaNanoRoleEnum.ROLE_CURATOR.toString()) &&
+						!aclSid.equalsIgnoreCase(CaNanoRoleEnum.ROLE_RESEARCHER.toString()) &&
+						!aclSid.equalsIgnoreCase(owner)) 
 						acl.deleteAce(i);
 					else
 						i++;
